Refactor automatic data selsctor
