+++ Tools/parametric_model/src/tools/__init__.py
@@ -4,3 +4,4 @@ from . import quat_utils
 from . import math_tools
 from .data_handler import DataHandler
 from .string_to_bool import string_to_bool
+from .automatic_data_selector import AutomaticDataSelector
